requirements.txt
streamlit
python-dotenv
requests
plotly
pandas
python-telegram-bot==20.3
supervisor
reportlab
